refactor api out of installable package
