+++ pyrox_api_service/__init__.py
@@ -1,0 +1,2 @@
+"""FastAPI reporting service package (repository-local, not shipped in wheel)."""
+
