+++ src/pyrox/__init__.py
@@ -1,16 +1,15 @@
+"""Public package exports for pyrox-client."""
+
 from .core import PyroxClient
-from .errors import PyroxError, RaceNotFound, AthleteNotFound
+from .errors import AthleteNotFound, PyroxError, RaceNotFound
 from .reporting import ReportingClient
-from .constants import *
 
 __version__ = "0.2.3"
 
 
-##  what is available to users
 __all__ = [
     "PyroxClient",
     "ReportingClient",
-    "build_athlete_options",
     "PyroxError",
     "RaceNotFound",
     "AthleteNotFound",
